requirements.txt
streamlit
pydantic
python-dotenv
baml-py